addTooltip {shinyBS}	R Documentation

addTooltip

Description

addTooltip is used within the Server logic of an app to add a tooltip to a Shiny input or output.

Usage

addTooltip(session, id, title, placement = "bottom", trigger = "hover",
  options = NULL)

Arguments

session
The session object passed to function given to shinyServer.

id
The id of the element to attach the tooltip to.

title
The content of the tooltip.

placement
Where the tooltip should appear relative to its target (top, bottom, left, or right). Defaults to "bottom".

trigger
What action should cause the tooltip to appear? (hover, focus, click, or manual). Defaults to "hover".

options
A named list of additional options to be set on the tooltip.

Details

See Tooltips_and_Popovers for more information about how to use addTooltip with the rest of the Tooltips_and_Popovers family.

Note

Run bsExample("Tooltips_and_Popovers") for an example of addTooltip functionality.

See Also

Twitter Bootstrap 3

Other Tooltips_and_Popovers: Tooltips_and_Popovers; addPopover; bsPopover; bsTooltip; popify; removePopover; removeTooltip; tipify

[Package shinyBS version 0.61 Index]