bsAlert {shinyBS}	R Documentation

bsAlert

Description

bsAlert creates an anchor point in your UI definition. This anchor point is where alerts created in your Server logic will be displayed.

Usage

bsAlert(anchorId)

Arguments

anchorId
A unique id the identifies the anchor.

Details

See Alerts for more information about how to use bsAlert with the rest of the Alerts family.

Note

Run bsExample("Alerts") for an example of bsAlert functionality.

See Also

Twitter Bootstrap 3

Other Alerts: Alerts; closeAlert; createAlert

[Package shinyBS version 0.61 Index]