bsButton {shinyBS}	R Documentation

bsButton

Description

bsButton is used in your UI script to create customizable action and toggle buttons.

Usage

bsButton(inputId, label, icon = NULL, ..., style = "default",
  size = "default", type = "action", block = FALSE, disabled = FALSE,
  value = FALSE)

Arguments

inputId
Specifies the input slot that will be used to access the value.

label
The contents of the button or link–usually a text label, but you could also use any other HTML, like an image.

icon
An optional icon to appear on the button.

...
Named attributes to be applied to the button or link.

style
A Bootstrap style to apply to the button. (default, primary, success, info, warning, or danger)

size
The size of the button (extra-small, small, default, or large)

type
The type of button to create. (action or toggle)

block
logical Should the button take the full width of the parent element?

disabled
logical Should the button be disabled (un-clickable)?

value
logical If type = "toggle", the initial value of the button.

Details

See Buttons for more information about how to use bsButton with the rest of the Buttons family.

Note

Run bsExample("Buttons") for an example of bsButton functionality.

See Also

Twitter Bootstrap 3

Other Buttons: Buttons; updateButton

[Package shinyBS version 0.61 Index]