bsCollapse {shinyBS}	R Documentation

bsCollapse

Description

bsCollapse is used in your UI to create a collapse panel group. Use bsCollapsePanel to populate this object with panels.

Usage

bsCollapse(..., id = NULL, multiple = FALSE, open = NULL)

Arguments

id
Optional You can use input$id in your Server logic to determine which panels are open, and updateCollapse to open/close panels.

multiple
Can more than one panel be open at a time? Defaults to FALSE.

open
The value, (or if none was supplied, the title) of the panel(s) you want open on load.

...
bsCollapsePanel elements to include in the Collapse.

Details

See Collapses for more information about how to use bsCollapse with the rest of the Collapses family.

Note

Run bsExample("Collapses") for an example of bsCollapse functionality.

See Also

Twitter Bootstrap 3

Other Collapses: Collapses; bsCollapsePanel; updateCollapse

[Package shinyBS version 0.61 Index]