bsCollapsePanel {shinyBS}	R Documentation

bsCollapsePanel

Description

bsCollapsePanel creates individual panels within a bsCollapse object.

Usage

bsCollapsePanel(title, ..., value = title, style = NULL)

Arguments

title
The title to display at the top of the panel.

value
Optional The value to return when this panel is open. Defaults to title.

style
Optional A Bootstrap style to apply to the panel. (primary, danger, warning, info, or success)

...
UI elements to include within the panel.

Details

See Collapses for more information about how to use bsCollapsePanel with the rest of the Collapses family.

Note

Run bsExample("Collapses") for an example of bsCollapsePanel functionality.

See Also

Twitter Bootstrap 3

Other Collapses: Collapses; bsCollapse; updateCollapse

[Package shinyBS version 0.61 Index]