bsExample {shinyBS}	R Documentation

bsExample

Description

A function to view examples of shinyBS functionality. Will run the examples found in the examples sections of shinyBS documentation. Use this instead of example.

Usage

bsExample(family, display.mode = "showcase", ...)

Arguments

family
A shinyBS family name

display.mode
The display mode to use when running the example. See runApp.

...
Other parameters to pass to runApp.

Details

This function is just a wrapper for runApp that runs copies of the examples found in the family documention pages of shinyBS. By default, display.mode is set to showcase so you can see the code while the app is running.

Examples

## Not run: 
   bsExample("Alerts")
## End(Not run)

[Package shinyBS version 0.61 Index]