bsModal {shinyBS}	R Documentation

bsModal

Description

bsModal is used within the UI to create a modal window.

Usage

bsModal(
  id,
  title,
  trigger,
  ...,
  size,
  footer = NULL,
  close.button = TRUE,
  width = NULL
)

Arguments

id
A unique identifier for the modal window

title
The title to appear at the top of the modal

trigger
The id of a button or link that will open the modal.

...
UI elements to include within the modal

size
Optional What size should the modal be? (small or large)

footer
A list of shiny UI elements to be added to the footer of the modal.

close.button
Should a close button be added to the footer of the modal?

width
An optional width argument for the modal. Must include units. Only applied if size is missing.

Details

See Modals for more information about how to use bsModal with the rest of the Modals family.

Note

Run bsExample("Modals") for an example of bsModal functionality.

See Also

Twitter Bootstrap 3

Other Modals: Modals, toggleModal()

[Package shinyBS version 0.64.0 Index]