bsTypeahead {shinyBS}	R Documentation

bsTypeahead

Description

typeaheadInput creates a textinput with type ahead function buttons.

Usage

bsTypeahead(inputId, label, value = "", choices, items = 8, minLength = 1)

Arguments

inputId
Input variable to assign the control's value to

label
Display label for the control

value
Initial value

choices
Array of strings to match against. Can also be JavaScript function. Use htmlwidgets::JS() to indicate JavaScript. The length of choices must no exceed 131370.

items
The max number of items to display in the dropdown. Can also be set to 'all'

minLength
The minimum character length needed before triggering autocomplete suggestions. You can set it to 0 so suggestion are shown even when there is no text when lookup function is called.

See Also

updateTypeaheadInput

[Package shinyBS version 0.62 Index]