closeAlert {shinyBS}	R Documentation

closeAlert

Description

closeAlert is used within your Server logic to close an alert that you created with createAlert.

Usage

closeAlert(session, alertId)

Arguments

session
The session object passed to function given to shinyServer.

alertId
The id of the alert to be dismissed.

Details

See Alerts for more information about how to use closeAlert with the rest of the Alerts family.

Note

Run bsExample("Alerts") for an example of closeAlert functionality.

See Also

Twitter Bootstrap 3

Other Alerts: Alerts; bsAlert; createAlert

[Package shinyBS version 0.61 Index]