createAlert {shinyBS}	R Documentation

createAlert

Description

createAlert is used within the Server logic of your Shiny app to display an alert to the user.

Usage

createAlert(session, anchorId, alertId = NULL, title = NULL,
  content = NULL, style = NULL, dismiss = TRUE, append = TRUE)

Arguments

session
The session object passed to function given to shinyServer.

anchorId
The unique identifier of the anchor where the alert should be displayed.

alertId
Optional A unique identifier for the Alert.

title
Optional A title for the Alert.

content
The main body of the Alert. HTML tags are allowed.

style
A bootstrap style to apply. Defaults to info.

dismiss
logical Should the Alert be user dismissable? Defaults to TRUE.

append
logical Should the Alert be appended below existing Alerts? Default to TRUE.

Details

See Alerts for more information about how to use createAlert with the rest of the Alerts family.

Note

Run bsExample("Alerts") for an example of createAlert functionality.

See Also

Twitter Bootstrap 3

Other Alerts: Alerts; bsAlert; closeAlert

[Package shinyBS version 0.61 Index]