popify {shinyBS}	R Documentation

popify

Description

popify can be wrapped around any shiny UI element to add a popover to the wrapped element. This should be a safer way to add popovers to elements created with renderUI.

Usage

popify(el, title, content, placement = "bottom", trigger = "hover",
  options = NULL)

Arguments

el
A shiny UI element.

title
The title of the popover.

content
The main content of the popover.

placement
Where the popover should appear relative to its target (top, bottom, left, or right). Defaults to "bottom".

trigger
What action should cause the popover to appear? (hover, focus, click, or manual). Defaults to "hover".

options
A named list of additional options to be set on the popover.

Details

See Tooltips_and_Popovers for more information about how to use popify with the rest of the Tooltips_and_Popovers family.

Note

Run bsExample("Tooltips_and_Popovers") for an example of popify functionality.

See Also

Twitter Bootstrap 3

Other Tooltips_and_Popovers: Tooltips_and_Popovers; addPopover; addTooltip; bsPopover; bsTooltip; removePopover; removeTooltip; tipify

[Package shinyBS version 0.61 Index]