removePopover {shinyBS}	R Documentation

removePopover

Description

removePopover is used within the Server logic of an app to remove an existing popover from a Shiny input or output.

Usage

removePopover(session, id)

Arguments

session
The session object passed to function given to shinyServer.

id
The id of the element to remove the popover from.

Details

See Tooltips_and_Popovers for more information about how to use removePopover with the rest of the Tooltips_and_Popovers family.

Note

Run bsExample("Tooltips_and_Popovers") for an example of removePopover functionality.

See Also

Twitter Bootstrap 3

Other Tooltips_and_Popovers: Tooltips_and_Popovers; addPopover; addTooltip; bsPopover; bsTooltip; popify; removeTooltip; tipify

[Package shinyBS version 0.61 Index]