removeTooltip {shinyBS}	R Documentation

removeTooltip

Description

removeTooltip is used within the Server logic of an app to remove an existing tooltip from a Shiny input or output.

Usage

removeTooltip(session, id)

Arguments

session
The session object passed to function given to shinyServer.

id
The id of the element to remove the tooltip from.

Details

See Tooltips_and_Popovers for more information about how to use removeTooltip with the rest of the Tooltips_and_Popovers family.

Note

Run bsExample("Tooltips_and_Popovers") for an example of removeTooltip functionality.

See Also

Twitter Bootstrap 3

Other Tooltips_and_Popovers: Tooltips_and_Popovers, addPopover(), addTooltip(), bsPopover(), bsTooltip(), popify(), removePopover(), tipify()

[Package shinyBS version 0.64.0 Index]