toggleModal {shinyBS}	R Documentation

toggleModal

Description

toggleModal is used within your Server logic to open or close a modal window.

Usage

toggleModal(session, modalId, toggle = "toggle")

Arguments

session
The session object passed to function given to shinyServer.

modalId
The id of the modal window you want to open/close

toggle
Should the modal window open, close, or toggle?

Details

See Modals for more information about how to use toggleModal with the rest of the Modals family.

Note

Run bsExample("Modals") for an example of toggleModal functionality.

See Also

Twitter Bootstrap 3

Other Modals: Modals; bsModal

[Package shinyBS version 0.61 Index]