updateButton {shinyBS}	R Documentation

updateButton

Description

updateButton is used in your Server logic to update the style or state of a button.

Usage

updateButton(session, inputId, label = NULL, icon = NULL, value = NULL,
  style = NULL, size = NULL, block = NULL, disabled = NULL)

Arguments

session
The session object passed to function given to shinyServer.

inputId
Specifies the input slot that will be used to access the value.

label
The contents of the button or link–usually a text label, but you could also use any other HTML, like an image.

icon
An optional icon to appear on the button.

value
logical If type = "toggle", the initial value of the button.

style
A Bootstrap style to apply to the button. (default, primary, success, info, warning, or danger)

size
The size of the button (extra-small, small, default, or large)

block
logical Should the button take the full width of the parent element?

disabled
logical Should the button be disabled (un-clickable)?

Details

Because of the way it is coded, updateButton may work on buttons not created by bsButton such as submitButton.

See Buttons for more information about how to use updateButton with the rest of the Buttons family.

Note

Run bsExample("Buttons") for an example of updateButton functionality.

See Also

Twitter Bootstrap 3

Other Buttons: Buttons; bsButton

[Package shinyBS version 0.61 Index]