updateCollapse {shinyBS}	R Documentation

updateCollapse

Description

updateCollapse is used within the Server logic of your Shiny app to modify a Collapse after load.

Usage

updateCollapse(session, id, open = NULL, close = NULL, style = NULL)

Arguments

session
The session object passed to function given to shinyServer.

id
The id of the Collapse object you want to change.

open
A vector of value (or title if no value was provided) values identifying the panels you want to open.

close
A vector of value (or title if no value was provided) values identifying the panels you want to close.

style
A named list of Bootstrap styles (primary, danger, info, warning, success, or default). The names should correspond to the value (or title if no value was provided) of the bsCollapsePanel you want to change.

Details

See Collapses for more information about how to use updateCollapse with the rest of the Collapses family.

Note

Run bsExample("Collapses") for an example of updateCollapse functionality.

See Also

Twitter Bootstrap 3

Other Collapses: Collapses; bsCollapsePanel; bsCollapse

[Package shinyBS version 0.61 Index]