bsModal {shinyBS}	R Documentation

bsModal

Description

bsModal is used within the UI to create a modal window.

Usage

bsModal(id, title, trigger, ..., size)

Arguments

id
A unique identifier for the modal window

title
The title to appear at the top of the modal

trigger
The id of a button or link that will open the modal.

size
Optional What size should the modal be? (small or large)

...
UI elements to include within the modal

Details

See Modals for more information about how to use bsModal with the rest of the Modals family.

Note

Run bsExample("Modals") for an example of bsModal functionality.

See Also

Twitter Bootstrap 3

Other Modals: Modals; toggleModal

[Package shinyBS version 0.61 Index]